Properly mark arbitrary JSON objects in OpenAPI
The OpenAPI schema's definitions of payload properties that allow arbitrary
JSON objects was inaccurate: it didn't include an `additionalProperties`
attribute to indicate that arbitrary keys and values were allowed.

This doesn't impact our existing client code but if we ever use the schema to
generate client code in statically-typed languages, the generated types would
be wrong.