SW-2775 Support localized enum names in search API
During code generation, generate properties files with human-readable (display)
names in English and gibberish for all the enum constants. These files are
considered build artifacts since they're generated from the database, but the
files for additional locales will go in `src/main/resources/i18n`.

Update the generated enum code so it loads localized names from the properties
files as needed. The English name continues to be available via the `displayName`
property, but it'll be removed in an upcoming change.

Update the search implementation so it uses the display names for the current
locale for matching, sorting, and rendering the query results.