Implement gibberish number formatting
To test localized number formats, the spec for the gibberish locale says it should
use ampersands as thousands separators and commas as decimal separators.

Implement a number formatter that uses those characters. Register it with Java's
localization system so call sites don't have to special-case the gibberish locale;
they can just call `NumberFormat` as usual.

Unfortunately, this requires adding a command-line argument to set a system
property when launching the server. Add it to the command line in the Docker image
and to the `generateOpenApiDocs` configuration, and add a check at start time to
tell people to include it in dev environments.