SW-2786 Make Messages text localizable (#917)
The `Messages` class is where we render human-readable text on the server side.
The intent was always that it'd be updated to return text in the user's language.

Move the English strings into a properties file and look them up based on the
current locale. The strings should be the same as before since we only have an
English properties file.

In addition, render the numeric parts of seed quantities using the number
formatter for the user's locale. This doesn't currently support the number
formatting rules for the gibberish locale; it just uses the built-in formatting
rules from the Java standard library.