SW-2800 Sort text search fields using locale
When the user sorts search results by a text field, we need to use the locale's
collation (sorting) rules or the results will potentially be in the wrong order.
PostgreSQL has built-in collation rules for a variety of languages; update the
search code to specify which rules to use for each query.