SW-2797 Localize country names (#925)
Make the search API return localized names for countries and country subdivisions.
The frontend uses the search API to get the list of available countries.

English country names are stored in the database, but they aren't turned into
Kotlin enums at build time, so the existing functionality for returning localized
strings for enum search fields won't work. But we still want to look up localized
names from resource bundles and generate the English and gibberish bundles from
the database, same as we do for enum tables.

To address this, introduce a new search field type, `LocalizedTextField`, that
works roughly the same way as `EnumField` but looks values up in resource bundles
explicitly rather than relying on the helper functions in
`EnumFromReferenceTable`.