SW-2727 Calculate total accession withdrawn quantities
Keep track of the total amount withdrawn from each accession, and expose the
totals as search fields.

Add a migration to add the totals to existing accessions. Since this isn't the
first time we've needed a "go through all the accessions one by one and compute
something" migration, add a helper method to `AccessionStore` to handle the
iteration logic. Once this migration is done, we'll remove
`MigrateWithdrawnTotals` but we can keep the helper method.