Clean up conditional bean declarations
To stop database-related code from running at start time when there's no database
(e.g., when generating OpenAPI docs) we used a Spring Boot annotation that checked
whether the `terraware.dailyTasks.enabled` property was set to `false`. That
works, but it's both verbose and misleading: we had the annotation on classes
that weren't actually daily tasks at all.

Define two new annotations that make the intent clearer:

`@DisableIfNoDatabase` does what the name implies: disables the bean if there
isn't a database configured, regardless of whether or not daily tasks are
enabled.

`@DailyTask` encapsulates the configuration check and is used on actual daily
tasks. It implies `@DisableIfNoDatabase`.