SW-2861 Generate Role enum from database
In anticipation of possibly supporting user-defined roles in the future, we had
been manually maintaining the `Role` enum to make it easier to replace with a
dynamic implementation later. We're no longer in imminent danger of implementing
user-defined roles, but the one-off implementaton wasn't hurting anything, so
we left it alone.

But now, this special-case implementation means we don't get localized role
names in search results. Rather than adding more special-case code to address that
problem, convert `Role` to a code-generated enum. This simplifies working with
role columns in the database, since conversion between the integer IDs and the
enum is now done by a code-generated converter class.

`Role` was the only reason `MappedField` existed, so remove it as well.