SW-2801 Use locale number formatting in search results (#937)
Currently, the search API renders numeric values with no thousands separator and
a period as the decimal separator. That's fine in English, but will be wrong in
locales that use different decimal separators.

Add locale-specific number formatting to all the numeric search fields, both at
display time and in exported CSVs. Formatted numbers are accepted as search
criteria as well.

Though this change is meant to support non-English locales, it also affects
English users: we will now show `1,234` instead of `1234` in search results.