SW-2904 Support CSV export of nested search fields
Commit 07039f0e added the ability to import a species list CSV file with multiple
ecosystem types separated by line breaks.

Allow exporting the species list in the same format.

The actual change here is more general: it removes the nested-fields restriction
on CSV export of search results. Previously, only flattened fields could be
exported because there is no way to represent tree structures in a CSV file.
That's still the case, even though we now allow nested fields to be exported;
if there are multiple levels of nesting, or if the values contain line breaks,
the results will be ambiguous. But for simple "list of enumerated values" cases
like species ecosystem type, there's no ambiguity.