SW-2896 Support custom storage location names (#952)
Update the API to support user-supplied storage location names:

* `GET /api/v1/seedbank/values/storageLocations` includes storage location IDs.

* The `facilities.storageLocations` search table now includes an `id` field.

* A new set of endpoints under `/api/v1/seedbank/storageLocations` allows
  creating, reading, updating, and deleting storage locations.

* `POST /api/v1/facilities` has a new optional field to set the list of storage
  location names to create.

* The system continues to create default storage locations if the client doesn't
  specify a list of names to create. The default storage location names are now
  localizable and are based on the locale of the admin who creates the seed bank.

Names are required to be unique within a facility.

For now, the accession API continues to use storage location names rather than IDs.

We're about to remove the concept of storage conditions from the system, so the
new APIs only require (and return) names and IDs. For now, if you use the existing
APIs to query storage locations that were created with the new APIs, they will
have "Freezer" as the storage condition, but that will be removed once the client
code no longer refers to storage conditions.