SW-2935 Support raw search fields (#964)
To allow clients to more easily parse specific values in search results, add
support for "raw" variants of fields whose values can vary based on locale.

Specify that you want to use raw values rather than localized ones by
appending the string `(raw)` to the name of a search field. You can use raw
fields both for input (search criteria) and for output (list of fields to return).

These are treated as separate search fields with separate names. It is fully
supported to include both the raw and localized versions of a field in the same
search request, and indeed, that's expected to be a common usage pattern whenever
a client needs to both display a value to a user and use it internally for
business logic.