SW-2908 Allow deleting seedling batches with photos (#969)
When you delete a seedling batch that is associated with a withdrawal, and the
withdrawal isn't also associated with other batches, the withdrawal is
automatically deleted as well. This was failing if the withdrawal had photos.

Add logic to delete the photos (including the files in the file store) before
attempting to delete the withdrawal from the database.