SW-2968 Fix Gradle warnings on generateOpenApiDocs (#970)
The Gradle `generateOpenApiDocs` task produced a bunch of warning messages. These
were due to https://github.com/springdoc/springdoc-openapi-gradle-plugin/issues/100
but luckily there's a workaround.

Also upgrade to the latest version of that plugin so we can remove Spring Boot
Devtools from  the classpath of the server that gets started by that task; it
spits out log messages at startup which make the build noisier.