SW-2877 API to edit partner reports (#966)
Add a set of endpoints under `/api/v1/reports` to manage quarterly reports from 
partners. Partners can read and update reports, but can't create or submit them
(creation and submission will be implemented separately).

To enforce the requirement that only one user at a time can edit a report, 
reports must be locked before they can be updated. (Reading a locked report is
allowed.) The basic workflow will look something like this:

1. `GET /api/v1/reports?organizationId=123` to list the available reports. The 
   search API can also be used; see below.
2. `GET /api/v1/reports/456` to fetch a report for viewing. If the report is
   still in progress, this will include the latest values for all the system-
   generated fields in the report.
3. `POST /api/v1/reports/456/lock` to lock a report for editing. This will 
   return an error if the report is already locked by another user.
4. (Optionally) `POST /api/v1/reports/456/lock/force` to steal the lock from 
   another user, presumably after asking for confirmation in the UI.
5. `PUT /api/v1/reports/456` to save the user's edits. This doesn't release the 
   lock, so it can be called multiple times to incrementally save work in
   progress.
6. `POST /api/v1/reports/456/unlock` to release the lock.

Since we expect the report format to evolve over time but we want historical 
reports to be treated as immutable, the report body format is versioned. 
In-progress reports are always upgraded to the latest version, but submitted
reports are always returned in the format that was used when they were
submitted.

To allow local testing, the organization page in the admin UI has a button to 
create a report for the previous calendar quarter. It's only available to super
admins.

Internally, the report is represented as a set of version-independent metadata
fields and a variable-format body. The body is stored as JSON in the database.

Also covers SW-2879, SW-2880, and SW-2881.