SW-2987 Rename photos to files
In preparation for adding the ability to attach non-photo files to reports, rename
the `photos` table to `files` and update various class and parameter names to
reflect the name change. The table itself is already generic; it doesn't have any
information that only applies to photos.

Child tables such as `accession_photos` keep their existing names because they
only hold information about photos, not about arbitrary files.

There are no changes to the API here, and there should be no user-visible behavior
change.

The majority of these changes were made using IntelliJ refactoring operations, but
there were manual edits in a few places to update comments and to keep lists of
names in alphabetical order.

DEPLOYMENT NOTE: If someone uploads a photo while this is in the middle of being
deployed, the upload will fail if their request happens to be routed to an instance
running the older code.