SW-2991 Add internal tagging of organizations
Our internal administrators need to be able to tell the system which
organizations are expected to submit quarterly reports. In addition, we've
gotten a steady stream of requests to run ad-hoc reporting queries on specific
subsets of organizations, or to exclude certain organizations from our
aggregated totals.

To solve both problems, add the concept of "internal tags" to the system.
Internal tags are, as the name suggests, only for internal use; they are not
exposed to end users in any way, and are only visible to super admins.

Define an initial fixed set of internal tags to cover some known use cases:
`Reporter` is for organizations that need to submit reports, `Internal` is for
our own internal projects, and `Testing` is for dummy organizations we use to
test the app.

In addition to these system-defined internal tags, add the ability to define
new ones via the admin UI. The admin UI also allows assigning internal tags to
organizations. These features are only available to super-admin users.

Currently, internal tags aren't used for anything by the system; subsequent
changes will introduce tag-dependent behavior. But analytics queries can
start using them right away.

This new feature uses the term "internal tag" rather than just "tag" because at
various times we have discussed adding a user-facing tagging feature. If we
implement that at some point, we'll want a clear distinction between internal
and user-facing tags.