SW-2987 Add endpoints to attach files to reports (#986)
Annual reports need to be accompanied by budget spreadsheet files, and it's likely
that in the future, we'll allow users to upload other kinds of supporting files as
well. Add endpoints to manage arbitrary files attached to reports.

These behave very similarly to the photo endpoints, but with some key differences:

* Thumbnails aren't supported.
* Files don't have captions, but their filenames are preserved.
* The GET endpoint to retrieve a file sets the `Content-Disposition` header so the
  files will be treated as downloads rather than inline content by browsers.

Internally, these non-photo files are tracked in the existing `files` table, but
there is a new linking table `report_files` to associate them with reports. This
adds some code duplication, but means that files can't be accidentally treated as
photos or vice versa since the two have different behavior.