SW-3000 Render reports as barebones HTML
In preparation for exporting submitted reports to Google Drive, add logic to
render report data as simple HTML, and allow the rendered reports to be viewed
using the admin UI.

The HTML includes references to photos and files. These are assumed to live in
the same directory as the exported HTML file. The admin UI includes endpoints
that serve the correct contents for those filenames.