SW-2992 Automatically create reports (#991)
Once a day, check whether or not there are organizations with the `Reporter`
internal tag that don't yet have a report for the previous calendar quarter. If
there are any, create reports for them.

For ad-hoc testing in dev environments, advancing the test clock using the admin
UI will cause daily tasks to run.

This requires a minor change to the permission logic in `OrganizationStore`,
which previously didn't allow the system user to fetch organization details
because the system user doesn't support enumerating its organization and
facility roles.