SW-3038 Add settings link to notification email (#998)
Include a link to the My Account page in the content areas of all our HTML
notification email messages, and to the footer area of our plaintext ones.

Since the requirements call for the link to be embedded in a sentence, add a
`link()` function email templates can use to render a localized string that
contains link text delimited by square brackets. This is the same approach we
use for link text in terraware-web and should make it easier to translate the
email messages.