Move ID and storage URL into FileMetadata
In preparation for implementing Google Drive export of report data, add fields for
the file ID and storage URL to `FileMetadata`. Since we use that class both for
new file uploads and for fetching existing files, introduce typealiases to make
the new fields null or non-nullable using the same scheme we use for other model
classes.

Remove redundant parameters and fields in places where we already have a
`FileMetadata` with the required information.

No change in client-visible behavior.