Add modified and submitted times/users to report API (#1007)
The designs for the reporting UI call for showing the names of the users who last
edited a report and who submitted a report. Add username, user ID, and timestamp
fields for modification and submission, in addition to the existing trio of fields
indicating who locked the report and when.