SW-2388 Add instructions to species CSV template (#1008)
Include lists of supported enum values and other guidance in the template for
importing species list CSV files.

Previously, per the initial requirements, we were checking for exact matches of
the header column names, but this was causing problems for users, and with the
instructions included, it's now much more likely that people will upload otherwise
valid CSV files where the header row isn't byte-identical to our template. Remove
the strict checking in favor of just checking that the number of columns in the
header row is correct.

Add test cases to ensure that localized templates work as intended.

Also partially addresses SW-3071.