SW-3070 Add ExactOrFuzzy search type (#1009)
Allow clients to specify whether or not they want the "try an exact search
first, then fall back to fuzzy if there were no matches" behavior introduced in
commit 64c854c. After getting user feedback, it turns out we only want that
behavior on a few specific search fields, not across the board.

Search field criteria with a type of `Fuzzy` are now always fuzzy, with no
exact-then-fuzzy fallback. The new type `ExactOrFuzzy` will return exact matches
if there are any, or fuzzy matches if there are no exact ones.