SW-2988 Upload submitted reports to Google Drive
When a report is submitted, export it to Google Drive along with its supporting
files. The report itself is created as a Google Docs document. A CSV file with
the aggregated report data is also created, and all photos and files are copied.

This requires some configuration:

* The `GOOGLE_APPLICATION_CREDENTIALS` environment variable must be set to the
  path of a Google client credentials configuration file. The file can be
  generated using the `gcloud` tool from the Google Cloud SDK.
* `terraware.report.export-enabled` must be true.
* `terraware.report.google-drive-id` must be set to the ID of a shared drive.
* `terraware.report.google-folder-id` may optionally be set to the ID of a folder
  within the shared drive; reports will be placed inside that folder.
* `terraware.report.google-email` may optionally be set to the email address of
  a user in our Google domain. If set, the server will impersonate that user when
  making API requests. This is needed for access to non-public shared drives.

If enabled, export happens automatically when a report is submitted. It can also
be triggered manually from the admin UI, where an "Export to Google Drive" button
will appear for each report on the organization details page.