SW-3203 Limit concurrent thumbnail generation
Loading too many photos into memory at once to generate thumbnails exhausts the
server's heap space. Add a semaphore to limit the number of photos we can resize
at the same time on a given server instance.

This initial implementation is very naive and doesn't take into account the fact
that memory usage varies depending on the photo's dimensions, but should be
conservative enough to protect us from out-of-memory errors if a client requests
dozens of thumbnails of newly-uploaded images in rapid succession.