Remove custom OpenAPI sorting
The OpenAPI library now has support for sorting schema elements automatically,
so use that instead of explicitly sorting everything ourselves.

Add tags to a couple controllers that were missing them so that they get sorted
into the proper positions in the Swagger UI, and add a test to catch any future
cases where we forget to add tags.