Remove redundant @RequestParam arguments (#1026)
Spring uses the controller argument names as the names of query string parameters
by default, so there's no need to specify them explicitly. Spring also infers
whether or not a query string parameter is required based on the nullability of
the argument's type, so there's no need to explicitly specify that in the
annotation either.

There's no change in behavior or the OpenAPI schema here.