SW-3236 Add Subzones and Monitoring Plots (#1028)
To accommodate upcoming changes to tracking, we had to update the database
to have another level of spatial hierarchy. Previously, there was planting
sites > planting zones > plots, but we need to change plots to planting
subzones, and add a level below plots called "monitoring plots". This
commit adds a database migration which performs the rename as well as adds
a new table for monitoring plots. We cleaned up the codebase for the renames
and added monitoring plots to existing unit tests.

The payload and search tables were also updated, so this is a breaking
change for the existing frontend. It should not be merged until those
necessary frontend changes are also ready to commit.