Move English enum display names to source code (#1029)
When we return enum values that are going to be shown to the user, we look up
localized display names from a properties file.

Previously, the properties files with the English display names were generated as
part of the build process. But that approach makes it difficult to include
additional notes to help translators pick the right translations.

Move the generated files into `src` and remove the logic that generates them from
the reference tables. Add a test to ensure that we don't forget to keep the files
up to date as the data model evolves.