SW-3183 Include Withdrawn Quantity in Seed Count
We want the total seed count for a seedbank to include the seed quantity
withdrawn from that seedbank as well as the seeds currently stored. This
result in the total number of seeds stored in the seedbank for all time
which is the value that is required for reports. We had to add the total
withdrawn seeds to the seedbank summay statistics to get this value.