Rename and sort English strings files
Phrase exports both the translated and English strings to GitHub, and it always
exports all the keys in alphabetical order.

Rename the English properties files to include a locale suffix to match the filenames
Phrase will update in its PRs, and sort the keys in `Messages_en.properties` to
match the version of the file that Phrase will export.