SW-3360 Disallow Lock/Unlock When Submitted (#1053)
We want to make sure the data in the database is consistent with its
submission status. Therefore, add a database migration which makes sure that
existing reports are in the submitted state when there is a submitted_by
user. It also adds a constraint that ensures that when there is such a user
the report must be submitted. And when the it is submitted there must be
a submitted_by user.

In the lock and unlock methods of ReportService, add a condition that the
report is not submitted and throw an exception otherwise.