SW-3352 MortalityRateInNursery Not Required (#1054)
A change in requirements means that for species in a planting site, the
mortality rate in nursery is no longer required. And the input for this
has been removed on the frontend. To allow a report to be validated,
remove the requirement. We also go through and remove the field from
the model and payloads.