Diff OpenAPI YAML in CI builds
Make it easy to see what changes a given PR would cause to the OpenAPI schema
by including a diff between the locally-generated schema and the staging one
in the GitHub Actions output.

Failure to fetch the schema from staging is logged but not treated as a build
failure (we don't want CI builds to fail if staging is down).