Use WGS84 coordinates in tracking tables (#1097)
Our API always uses WGS84 (degrees latitude and longitude) when it returns
geometry values to the client, since GeoJSON uses that coordinate system.  But
we've been transforming coordinates to "web Mercator" in the database at write
time, meaning we have to do on-the-fly coordinate system transformation every
time we read a geometry value.

The cost of those transformations is negligible for small amounts of data, but
becomes significant in the face of large planting sites with tens of thousands
of monitoring plots.

Change the code to store WGS84 coordinates in the database instead. Remove
the read-time transformations since they're no longer needed.