SW-3491 Support tracking v2 shapefile import (#1098)
Update the shapefile importer to support the tracking v2 hierarchy with subzones
and monitoring plots. Import of v1 shapefiles is no longer supported.

The top three levels of the hierarchy (site, zones, and subzones) are loaded from
shapefiles. Monitoring plots are generated programmatically.

The planting site page in the admin UI now shows a map view of the planting site.
It no longer shows a full list of all monitoring plots, just a count of them for
each subzone.

This also introduces a change to how the names of monitoring plots and subzones
are sorted. The names are arbitrary text values and are alphanumeric in the
initial shapefiles we'll be importing, but per the requirements doc, they're
supposed to be integers. We want to sort them in numeric order if the shapefile
was compliant with the spec, but still allow non-numeric names so we can import
the initial set of shapefiles. So we use a collation that detects numbers in
strings and sorts them numerically, while still sorting the non-numeric parts
alphabetically.

For now, we don't have any anonymized example shapefiles to use in automated
tests, so most of the validation tests are removed.