SW-3521 Generate permanent monitoring plot clusters (#1111)
Group monitoring plots into 2x2 clusters. If all four plots in a cluster fall
within a single subzone, that cluster is eligible to be chosen as a set of
permanent monitoring plots.

An eligible cluster is identified by the presence of a permanent monitoring plot
cluster number. All four plots in a cluster will have the same cluster number.
The cluster numbers are randomized at import time.

The admin UI is updated to demonstrate/test the selection process; currently it
is hardwired to simulate 6 clusters of permanent plots per planting zone and
highlight them on the map.