SW-3535 Add permissions for planting zones
In preparation for being able to configure the number of monitoring plots on a
per-planting-zone basis, add read and update permission logic for planting zones.
There are no create or delete permissions since planting zones are only created
via shapefile import, and are never deleted independently of the planting site.