SW-3535 Support per-planting-zone plot counts (#1115)
For each planting zone, keep track of how many permanent and temporary monitoring
plots will be needed for the next observation.

Ordinarily, these numbers will be calculated based on statistical information
about the planting zone, expressed as variance and error margin along with a value
that represents the desired confidence level (Student's t). The confidence level
is expected to stay the same across all planting sites, so it is initialized to
the correct value when a shapefile is imported. The other two values will be
supplied by the science team prior to the first observation of a planting site.

Those numbers are also stored on a per-planting-zone basis, and can be edited in
the admin UI. The admin UI uses those numbers to calculate the number of permanent
monitoring plots, which is then used to calculate the number of temporary plots.
These calculations are done client-side in the admin UI so as to allow the admin
to manually adjust the plot counts before saving if needed. The plot counts can
also be entered directly without using the formula at all.

The map in the admin UI shows the correct number of permanent clusters for each
planting zone rather than a hardwired number of them.