SW-3527 Allow Dashes in Scientific Names During Import (#1120)
Update the import validator to allow dashes in scientific names. We allow for
all three types of dashes that might be encountered, hyphen, en-, and em-.
Also update the tests to account for this change. When saving, all dash types
are normalized to hyphens to prevent confusion and unintentional duplicates.