SW-3509 Handle duplicate accession photo filenames
Initially, all the metadata about accession photos was stored in a single table
that had a unique constraint to enforce that a given accession could only have
one photo with a given filename. When the table was split up to add support for
nursery withdrawal photos, we lost that unique constraint and thus the server
started accepting uploads of accession photos with duplicate filenames.

When a client attempted to read one of those photos, the request would fail
because the server still assumed filenames were unique and a database query
that was expected to return a single row instead returned multiple rows.

As an immediate fix for the request failures, update the query to return the
most recent file with a given name.