Use separate map layer for permanent plots (#1123)
When showing the map of a planting site in the admin UI, we highlight permanent
monitoring plots. Previously, this was done by selecting the line color, but that
meant that on the grid of plots, only the left and bottom borders were colored
because the right and top ones were overwritten by other grid cells.

Add a separate layer for permanent plots so they show up more distinctly on the
map. Restructure the way the plot types are labeled in the GeoJSON data so that
it will be possible to do the same for temporary plots later.