Show area and reported plants in each subzone
Include the area in hectares and the number of reported plants in each subzone
on the admin UI's planting site page. These numbers will be used to calculate
reported planting density as well as to determine which subzones are eligible
for monitoring, so being able to view them easily is useful for sanity-checking.