SW-3070 Exact Search Performs Substring Search for Selected Fields
We want to have, for accession numbers and collection sites, among other,exact
search to return results with substring exact matches. This change updates the
TextField and UpperCaseTextField exact search conditions to account for this.
When exact substring matches are available, trigram (fuzzy) search should not
be in effect for exact-or-fuzzy search.