Suppress load balancer cookie warning
The Keycloak adapter uses an out-of-date version of the Apache HTTP client that
doesn't know how to deal with the expiration times on the cookies that are added
by the load balancer that sits in front of our Keycloak instance. It spits out
a warning log message every time the adapter talks to Keycloak, which happens
often enough to add significant log clutter.

The eventual fix will be to stop using the Keycloak adapter (which is deprecated)
but for now, configure the logger to filter out warnings from the class in
question.