SW-3601 Create observations from admin UI
Add an "Observations" section to the planting site page in the admin UI that lists
all the site's observations. If the user is a super admin, there is also a form to
create a new observation, with the start and end dates based on the site's
planting season end month by default.

Automatic scheduled creation of observations will come in a future change.

This change also adds the concept of an observation state. Currently, since the
only operation is "create," observations will always be in the initial "Upcoming"
state; future changes will add transitions to other states.