Use inserted test IDs as defaults
Make the insert helper functions in `DatabaseTest` remember the IDs of the rows
they've inserted, and use the most recently inserted ID as the default in places
where callers previously had to pass in IDs explicitly.

This will help us write more succinct tests, especially in cases where a complex
graph of entities has to be built.

Update one of the tests in `PlantingSiteStoreTest` to demonstrate how we can take
advantage of this change.