SW-3601 Start observations from admin UI (#1145)
* SW-3601 Start observations from admin UI

Implement the logic for assigning monitoring plots to an observation. In the
finished product, this will run automatically when the observation's start date
arrives, but for now super-admin users can trigger it manually via the admin UI
if the prerequisites are met (mainly: the start date has arrived and the site
has subzones with plants).

* Address review feedback