SW-3639 All subzones of permanent cluster must be planted (#1164)
Now that we allow monitoring plots to straddle subzone boundaries, it is also
possible for a permanent cluster to be made up of plots in multiple subzones.
The "only assign permanent clusters whose subzones are planted" logic thus needs
to be changed such that it requires _all_ the cluster's subzones to be planted.