SW-3641 Planting zone target planting densities
Track the target planting density (in plants per hectare) for each planting zone.
The value is imported from the shapefile if present; otherwise we use a default.
It can be updated using the admin UI.

Nothing uses this value yet, but it will be an input into planting progress
calculations.