Handle apostrophes in enum display names
The localized English name for `RecordedSpeciesCertainty.CantTell` was being
rendered as `Can''t Tell` because single quotes in properties files are always
escaped by Phrase and we weren't unescaping them.

Including an apostrophe in an enum's name in the reference table was generating
broken code because it isn't a valid character in a Kotlin symbol. Update the
code generator to strip non-alphanumeric characters out of enum values.