Accept canonical enum values in API payloads (#1169)
JSON deserialization of our reference table enums was expecting English
display names as values, rather than the non-localized names from the
`name` columns of the reference tables.

Historically, those two were the same since we didn't have any concept of
localizing an enum name. But now that we support localization, there is
a distinction between the enum string values in the API (which uses the
names from the database) and the human-readable values that get included
in exported CSV files and the like.

We were correctly using the names from the database in responses, and listing
those names as the acceptable values in the OpenAPI schema. But the actual
JSON deserialization was expecting the localized English names.

This caused the server to reject the value `CantTell` of the
`RecordedSpeciesCertainty` enum because its localized English name is "Can't
Tell". (Currently, that's the only enum value whose string representation
in the API differs from the localized English name.)