Remove facility scope from seedbank search API
`/api/v1/seedbank/values/all` supports listing the values for an organization or a
facility, but we no longer ever use the facility scope. Replace the scoping
implementation with a simpler one that only supports organizations.

The fact that organization ID is a required field in the request payload is now
part of the OpenAPI schema instead of a runtime check.